add babelsrc
