--- config/conf.json
@@ -1,3 +1,0 @@
-{
-  "name": "Node.js Homework Application"
-}
