--- models/Product.js
@@ -1,7 +1,0 @@
-class Product {
-    constructor() {
-        console.log(`Product module`);
-    }
-}
-
-exports.default = Product;
