--- models/User.js
@@ -1,7 +1,0 @@
-class User {
-    constructor() {
-        console.log("User module");
-    }
-}
-
-exports.default = User;
