--- models/models.js
@@ -1,4 +1,0 @@
-module.exports = {
-    User :  require('./User.js').default,
-    Product :  require('./Product.js').default,
-}
