+++ src/config/conf.json
@@ -1,0 +1,3 @@
+{
+  "name": "Node.js Homework Application"
+}
