+++ src/models/Product.js
@@ -1,0 +1,7 @@
+class Product {
+    constructor() {
+        console.log(`Product module`);
+    }
+}
+
+export default Product;
