+++ src/models/User.js
@@ -1,0 +1,7 @@
+class User {
+    constructor() {
+        console.log("User module");
+    }
+}
+
+export default User;
