+++ src/models/models.js
@@ -1,0 +1,4 @@
+module.exports = {
+    User :  require('./User.js').default,
+    Product :  require('./Product.js').default,
+}
